// src/views/PageLayoutView/component-library.element-plus.generated.ts
export interface ElementPlusAutoRenderBinding {
  type: string
  tag: string
  slotTextProp: string | null
  optionItemsProp: string | null
  tableColumnKeysProp: string | null
}
// ...
export const elementPlusAutoRenderBindingList: ElementPlusAutoRenderBinding[] = [
  {
    "type": "epButton",
    "tag": "el-button",
    "slotTextProp": "slotText",
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epLink",
    "tag": "el-link",
    "slotTextProp": "slotText",
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epInput",
    "tag": "el-input",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epInputNumber",
    "tag": "el-input-number",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epSelect",
    "tag": "el-select",
    "slotTextProp": null,
    "optionItemsProp": "optionItems",
    "tableColumnKeysProp": null
  },
  {
    "type": "epCascader",
    "tag": "el-cascader",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epDatePicker",
    "tag": "el-date-picker",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epTimePicker",
    "tag": "el-time-picker",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epTimeSelect",
    "tag": "el-time-select",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epSwitch",
    "tag": "el-switch",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epSlider",
    "tag": "el-slider",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epRate",
    "tag": "el-rate",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epColorPicker",
    "tag": "el-color-picker",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epCheckbox",
    "tag": "el-checkbox",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epRadio",
    "tag": "el-radio",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epTag",
    "tag": "el-tag",
    "slotTextProp": "slotText",
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epAlert",
    "tag": "el-alert",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epProgress",
    "tag": "el-progress",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epBadge",
    "tag": "el-badge",
    "slotTextProp": "slotText",
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epAvatar",
    "tag": "el-avatar",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epEmpty",
    "tag": "el-empty",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epResult",
    "tag": "el-result",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epStatistic",
    "tag": "el-statistic",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epPagination",
    "tag": "el-pagination",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epTable",
    "tag": "el-table",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": "columnKeys"
  },
  {
    "type": "epCard",
    "tag": "el-card",
    "slotTextProp": "slotText",
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epDivider",
    "tag": "el-divider",
    "slotTextProp": "slotText",
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  },
  {
    "type": "epSegmented",
    "tag": "el-segmented",
    "slotTextProp": null,
    "optionItemsProp": null,
    "tableColumnKeysProp": null
  }
] as ElementPlusAutoRenderBinding[]